Mssql: Remove `freetds` from supported driver list
There's no mention of this identifier on
https://www.php.net/manual/en/ref.pdo-dblib.connection.php.

I also found no mention of it at any time
(using the wayback machine)

So there's only Zend's v1 adapter which uses it as a fallback
for `sybase`, which implies it being only an alias. Since we
are comparing this with the result of `PDO::getAvailableDrivers()`
and don't use it as configuration option, I highly doubt we
retrieve aliases there.